move badger options
