+++ node/store_badger.go
@@ -41,3 +41,24 @@ func init() {
 	storeConstructors[BadgerStore] = constructor
 	storeConstructors[DefaultStore] = constructor
 }
+
+// WithBadgerInMemory sets the badger in memory option.
+func WithBadgerInMemory(enable bool) StoreOpt {
+	return func(o *StoreOptions) {
+		o.badgerInMemory = enable
+	}
+}
+
+// WithBadgerFileSize sets the badger value log file size.
+func WithBadgerFileSize(size int64) StoreOpt {
+	return func(o *StoreOptions) {
+		o.badgerFileSize = size
+	}
+}
+
+// WithBadgerEncryptionKey sets the badger encryption key.
+func WithBadgerEncryptionKey(encryptionKey []byte) StoreOpt {
+	return func(o *StoreOptions) {
+		o.badgerEncryptionKey = encryptionKey
+	}
+}
